+++ packages/common/src/capitalize.ts
@@ -1,5 +1,5 @@
 export const capitalizeFirstLetter = (text: string): string => {
-  if (text?.length > 0) {
+  if (text) {
     return text.charAt(0).toUpperCase() + text.slice(1)
   } else {
     return ''