+++ packages/state-manager/src/sagas/socket/startConnection/startConnection.saga.ts
@@ -42,6 +42,8 @@ import {
   type SavedOwnerCertificatePayload,
   SendUserCertificatePayload,
   type SendOwnerCertificatePayload,
+  SaveCSRPayload,
+  CommunityMetadata,
 } from '@quiet/types'
 
 const log = logger('socket')
@@ -86,7 +88,10 @@ export function subscribe(socket: Socket) {
     | ReturnType<typeof connectionActions.setTorBootstrapProcess>
     | ReturnType<typeof connectionActions.setTorConnectionProcess>
     | ReturnType<typeof connectionActions.torBootstrapped>
+    | ReturnType<typeof communitiesActions.clearInvitationCodes>
+    | ReturnType<typeof identityActions.saveUserCsr>
     | ReturnType<typeof connectionActions.setTorInitialized>
+    | ReturnType<typeof communitiesActions.saveCommunityMetadata>
   >(emit => {
     // UPDATE FOR APP
     socket.on(SocketActionTypes.TOR_INITIALIZED, () => {
@@ -158,10 +163,12 @@ export function subscribe(socket: Socket) {
 
     // Community
     socket.on(SocketActionTypes.NEW_COMMUNITY, (_payload: ResponseCreateCommunityPayload) => {
+      console.log('on SocketActionTypes.NEW_COMMUNITY')
       emit(identityActions.saveOwnerCertToDb())
       emit(publicChannelsActions.createGeneralChannel())
     })
     socket.on(SocketActionTypes.REGISTRAR, (payload: ResponseRegistrarPayload) => {
+      console.log('SocketActionTypes.REGISTRAR')
       log(SocketActionTypes.REGISTRAR, payload)
       emit(communitiesActions.responseRegistrar(payload))
       emit(networkActions.addInitializedRegistrar(payload.id))
@@ -176,8 +183,10 @@ export function subscribe(socket: Socket) {
     socket.on(SocketActionTypes.COMMUNITY, (payload: ResponseLaunchCommunityPayload) => {
       console.log('Hunting for heisenbug: Community event received in state-manager')
       emit(communitiesActions.launchRegistrar(payload.id))
+      emit(identityActions.saveUserCsr())
       emit(filesActions.checkForMissingFiles(payload.id))
       emit(networkActions.addInitializedCommunity(payload.id))
+      emit(communitiesActions.clearInvitationCodes())
     })
     // Errors
     socket.on(SocketActionTypes.ERROR, (payload: ErrorPayload) => {
@@ -194,9 +203,8 @@ export function subscribe(socket: Socket) {
       )
       emit(usersActions.responseSendCertificates(payload))
     })
-
     socket.on(SocketActionTypes.SEND_USER_CERTIFICATE, (payload: SendOwnerCertificatePayload) => {
-      console.log('user cert with owner cert', payload)
+      console.log('Received SEND_USER_CERTIFICATE', payload.communityId)
 
       emit(
         communitiesActions.addOwnerCertificate({
@@ -226,6 +234,7 @@ export function subscribe(socket: Socket) {
       emit(communitiesActions.launchCommunity(payload.communityId))
     })
     socket.on(SocketActionTypes.SAVED_OWNER_CERTIFICATE, (payload: SavedOwnerCertificatePayload) => {
+      console.log('Received SAVED_OWNER_CERTIFICATE', payload.communityId)
       emit(
         communitiesActions.addOwnerCertificate({
           communityId: payload.communityId,
@@ -245,6 +254,15 @@ export function subscribe(socket: Socket) {
         })
       )
       emit(identityActions.savedOwnerCertificate(payload.communityId))
+    })
+    socket.on(SocketActionTypes.SAVE_COMMUNITY_METADATA, (payload: CommunityMetadata) => {
+      console.log('SAVE COMMUNITY METADATA', payload)
+      emit(
+        communitiesActions.saveCommunityMetadata({
+          rootCa: payload.rootCa,
+          ownerCertificate: payload.ownerCertificate,
+        })
+      )
     })
     return () => undefined
   })