+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 [unreleased]
 
+* Replaced greying out inputs with splash screen on joining/creating screens.
+
+* Changed registration process - user connects to the libp2p network directly instead of using registrar. Invitation link format changed. User csr is now saved to database.
+
 * Fixed android stucking on username registration screen introduced in previous alpha.
 
 * Added creator username to initial channel message.
@@ -7,5 +11,9 @@
 * Fixed bug with changing joining community/create community screens with required field.
 
 * Fixed bug with displaying incorrect default settings tab.
+
+* Replaced source of publicKey in sendMessage saga to CSR
+
+* Labels for unregistered and duplicate usernames with modals
 
 * Fixed LoadingPanel useEffect bug.